feat: moved schema validation to service
